Bug fix.  For standalone plots/tables; on subsequent calls, all previous event callback instantiations would fire again, resulting in an increasing number of plots/tables being rendered.  Now, the callbacks are unregistered after they've been called.
Along with this, the behaviour to add to an existing base plot/table has (finally) been sorted out.  It is controllable by a property (append), which is initially set to false, causing a new plot/table to be rendered on each invocation by default.  Setting to true, will cause data to be added to the existing base plot/table.