The isInitLoad sentinel, which is used to indicate whether to handle an OL zoomend event, is now set true/false solely within SOS.Map._display().  This is required because when using a (non-standard) map size, for which maxResolution has to be set "auto", then no initial zoom takes place (as a result of map.zoomToMaxExtent()), whereas it did before with a standard (512x256) map size.  The upshoot of this, was that isInitLoad wasn't being cleared in SOS.MapSet.mapZoomEndHandler(), meaning that for the first genuine user zoom to, say, Halley, on the world map, it didn't auto switch to the South map, as it should, because isInitLoad was still set true.
Although this means that a later component (MapSet) has influenced an earlier component (Map), which I was trying to avoid, it actually makes things much cleaner & simpler.

Also, we now pass the OL 'extra' optional arguments when initialising the baseLayer, to allow for setting maxResolution = 'auto' as required for non-standard map sizes.