Mods. to datetime handling.
A new function formats the datetime strings used in a get observations request, making it easier to override on a per-application basis.  By default, this function formats as ISO8601 without the subsecond component, for broadest applicability.

Also, the datetime parsing utility functions now parse timezone offsets, and shift to UTC.