Optionally use a FOI Id to identify a single station in a multi-station offering.
For those data models that have multiple station data per offering, setting this option instructs the app to include the Id of the selected FOI (from the map) in a get observation request.  This then filters the offering result so that it returns data only for the selected station.  This is optional, because some applications may want all stations' data, for example, to plot a separate curve for each station.