Minor mods.  SOS.App datepickers are corrected to UTC.
The SOS.App uses jQuery datepickers to select a date range,  then adds time so that the interval starts at midnight of the selected start date, & ends at 23:59:59.999 of the selected end date.  A datepicker returns a javascript date object, with the client's timezone.  These mods.  correct the given date object to be UTC.