All standalone components can now optionally use a FOI Id to identify a single station in a multi-station offering.
For those data models that have multiple station data per offering, explicitly setting the foiId property instructs the component (SOS.Plot, SOS.Table, SOS.Plot.Rose etc.) to include the FOI Id in a get observation request.  This then filters the offering result so that it returns data only for the selected station.  If foiId is set null (the default) then data for all the offering's stations are returned.  This is optional, because some applications may want all stations' data, for example, to plot a separate curve for each station.