Changed SOS.App default option for including FOI Id in getObservations.
An option was introduced to allow conditional inclusion of a FOI's Id (when selected from a map) in a getObservations call from a SOS.App object (see commit d8891ba8f343427e75cb8da25e45de72e502bcda).

Originally, it defaulted to false.  This was so that the app would work as previously.  However, having it true by default means that for a single-station offering, it essentially works as before anyway (as there is only one FOI), whereas for a multi-station offering, it will only plot/tabulate data for that selected station; which is almost certainly what you would always want to do anyway.  As before, setting this to false will override this behaviour.