Show data for all stations of a multi-station offering, if no FOI Id is specified.
The main UI data display components (SOS.Plot and SOS.Table) will now show data for all stations of a multi-station offering, if no FOI Id is specified.  For example, say you have an offering with three stations, and you wish to plot them all on the same plot.  Then omitting a FOI Id, and setting the forceSingleAxis plot option, will plot each station's data as a separate curve on a single plot.

Along with these mods., a label template system has been introduced, to allow specifying which metadata to include in the labels of a SOS.Plot or SOS.Table (or derived display objects).  See the labelTemplates plot/table option for details.