General mechanism to configure the components of a SOS.App.
Any properties set on sub-objects of SOS.App.config.app.options.components, (via the function setAppComponentsOptions()) will be passed down to the corresponding component on instantiation.

For example:

app.setAppComponentsOptions({
  plot: {
    plot: {options: {zoom: {interactive: false}}},
    format: {value: {digits: 1}}
  },
  table: {
    table: {options: {scrollable: true}}
  }
});

will turn off mouse-scrollwheel zooming for the app's plot, and show the values displayed in the plot's tooltips to one decimal place.  Furthermore, it will ensure that the app's table is scrollable.