Mods. to SOS.Map to make it easier to customise the map.
SOS.Map now has an option: useOsm, to use OpenStreetMap as the map's base layer, rather than a WMS.  Along with this, it's now easier to set the map's initial view, using either the centre and zoom options, or the params.restrictedExtent option.

These mods., along with those of the previous commit (921b7d7369ecc77160e30f91af6219959124b56c - general mechanism for configuring a SOS.App's components) now mean it is far easier to customise the map component of a SOS.App.

For example:

app.setAppComponentsOptions({
  // Setup map component to use OpenStreetMap
  map: {
    map: {
      options: {
        defaultProjection: new OpenLayers.Projection("EPSG:4326"),
        centre: new OpenLayers.LonLat(146.48, -41.7).transform(new OpenLayers.Projection("EPSG:4326"), new OpenLayers.Projection("EPSG:900913")),
        zoom: 8,
        params: {
          projection: "EPSG:900913",
        }
      }
    },
    baseLayer: {options: {useOsm: true}},

    // This SOS has its FOI coordinates as lon/lat
    featureOfInterestLayer: {options: {crs: {format: {xy: true}}}}
  }
});

will set the baselayer to use OpenStreetMap and centre the map on Northern Tasmania.  Furthermore, this SOS instance's FOI coordinates are in lon/lat, so we inform OpenLayers of this fact.