Bugfix.  https://github.com/52North/sos-js/issues/17.
The temporal coverage for the whole SOS instance is often present in the service's capabilities document, and is used to populate the right-hand Metadata info box in the SOS.App.  This mod. improves the robustness of checking that this temporal coverage exists in the capabilities document, before trying to use it.