Added foiFilter function to SOS.Map.
By default, this function simply passes the FOIs array from the SOS
as-is to the map's FOI layer.  However, this function can be overridden
to selectively filter which FOIs to display on the map.

For example:

app.setAppComponentsOptions({
  map: {
    featureOfInterestLayer: {
      options: {
        foiFilter: function(fois) {
          var filteredFois = [];

          for(var i = 0, len = fois.length; i < len; i++) {
            if(fois[i] == "foi_123" || fois[i] == "foi_456") {
              continue;
            } else {
              filteredFois.push(fois[i]);
            }
          }

          return filteredFois;
        }
      }
    }
  }
});

will remove the two FOIs with IDs foi_123 and foi_456 from the list of
FOIs obtained from the SOS, and so these two FOIs will not be displayed
on the map.