SOS.Map now supports multiple FOI layers.
The SOS.Map component now has a featureOfInterestLayers property, which is an array.  By default, the first element of this array is the default FOI layer, and a reference to this is held in the featureOfInterestLayer property, for backwards-compatibility.

To add a new FOI layer, simply push a layer config describing the layer to the featureOfInterestLayers array.  See the default FOI layer in the SOS.Map component for the structure of a layer config.

For example, to add a new layer that will show ships separately from fixed stations, we first define the layer config object:

var basShipsFeatureOfInterestLayer = {
  object: null,
  id: "sosMapBasShipsFeatureOfInterestLayer",
  options: {
    label: "BAS Ships",
    pointStyle: new OpenLayers.Style(basPointStyleOptions),
    crs: {format: {xy: null}},
    displayLatestObservations: false,
    // Only show BAS ship FOIs on this layer
    foiFilter: this.showBasShipFois
  }
};

Note that we define a custom FOI filter function (this.showBasShipFois), that will only show our ships on this layer.  See the previous commit (a3c6c4b899c2080c4f04c702aab6ff42b53762be) for details of this FOI filter function.

We then add the layer config to the array:

map.config.featureOfInterestLayers.push(basShipsFeatureOfInterestLayer);

Then when the map is subsequently displayed:

map.display();

it will contain two layers; the default layer, and this new ships layer.

Note also that the default layer can be customised, either via a call to map.setFeatureOfInterestLayerOptions(), or by completely replacing the zeroth element of the featureOfInterestLayers array.