Allow a FOI layer's initial visibility to be set.
A FOI layer's initial visibility state can be set on instantiation of the layer.  It defaults to true, which means it will be loaded on map load (if it's in range), and checked in the layerswitcher.  If false, it will not be loaded on map load, but will be listed in the layerswitcher, unchecked.  Checking the layer will load it on the map.

This visibility property can be changed via the setFeatureOfInterestLayerOptions function.  For example:

map.setFeatureOfInterestLayerOptions({visibility: false});

See also previous commits (8939afe316f0764a90a1cbdbab29c52cb7bb89d2 and a3c6c4b899c2080c4f04c702aab6ff42b53762be) for details of creating additional FOI layers.