General mechanism for passing FOI layer options on instantiation.
Any options set in the params member will be passed as-is to OpenLayers when constructing the FOI layer.

For example, to set a custom event handler on the layer:

map.setFeatureOfInterestLayerOptions({
  params: {
    eventListeners: {"loadstart": foiLayerLoadStartHandler}
  }
});

Unfortunately, this can't be used to set the visible state of a layer (via the visibility property), as OpenLayers only changes the visibility of a layer via a call to the layer's setVisibility() function.  See the previous commit (d23bd4135ad318bf65a230e636396935dafd641a) for details.