Utility object to display busy indicator.
New SOS.Ui utility object SOS.Ui.BusyIndicator, to display a pulsing busy
indicator.  Typically used to inform the user that an asynchronous
process is occurring, for example, a component is fetching data from the
service endpoint.  Takes options upon instantiation, probably the most
important of which is a parent container for the busy indicator box to be
appended to.

Example:

map.setFeatureOfInterestLayerOptions({
  params: {
    eventListeners: {
      loadstart: function() {
        this.busyIndicator = SOS.Ui.BusyIndicator;
        this.busyIndicator.init({
          boxParent: ".sos-container-centre",
          boxClass: "sos-busy-indicator-box sos-centred-busy-indicator"
        });
      },
      loadend: function() {
        if(this.busyIndicator) {
          this.busyIndicator.stop();
        }
      },
      scope: this
    }
  }
});