Bugfix for SOS.Map.constructFeatureOfInterestLayerLoadHandlers().
The SOS.Map.constructFeatureOfInterestLayerLoadHandlers() function has been simplified.  Before, if the showBusyIndicatorOnLoad option was set, and the user had specified a custom loadstart/loadend handler, the function would attempt to incorporate the user's custom loadstart/loadend handler, as well as a busy indicator.  Now it overwrites the user's custom handler, and so just calls the busy indicator.  See previous commit (185e304f1f3d3135aa2514db64379899a0533996).

The reason for this is that, in trying to incorporate both handlers, it creates temporary functions on the given scope object, which works OK.  However, in a long-lived app, where the map is redrawn many times, this results in the previously combined functions being incorporated again, eventually resulting in a fatal 'Too much recursion' bug.

All this means though, is that if the user wants this dual functionality, then they have to turn off the automatic creation of a busy indicator by setting showBusyIndicatorOnLoad = false, and then create the busy indicator start/end calls in their custom handler(s).

For example:

customFoiLayer.options = {
  showBusyIndicatorOnLoad: false,
  params: {
    eventListeners: {
      loadstart: this.customLoadstartHandler,
      loadend: this.constructBusyIndicatorEndHandler(),
      scope: this
    }
  }
};

customLoadstartHandler = function(evt) {
  // Start the busy indicator via a temporary function, then remove
  this._f = this.constructBusyIndicatorStartHandler("#id");
  this._f(evt);
  delete this._f;
  // Do other things whilst this layer is loading...
}