Mods. to Stuve plot.  It now includes altitude.
The Stuve plots include altitude as a right-hand vertical axis.  As the primary vertical axis (pressure) is non-linear, it will only show certain altitude values that fall on a pressure grid line (by default, every 50 hPa).  Altitude is also included in the tool-tip when hovering over the plot, as well as the data series that can be requested from the plot.