SOS.Table can optionally merge series into a single table.
When multiple observed properties are added to a SOS.Table, they are independent tables (i.e., each have time and value columns).  Setting the mergeSeries.active option to true, will mean that when the table is rendered, all observed properties will be merged into a single table.  This is done by finding the unique set of all time points, and then matching each observed property value (or the missing option - default: "") to this time point.  Essentially, it's a left outer join.

Bear in mind that this is done client-side, and so for large series can incur significant overhead.  For this reason, and for backwards compatibility, this option is turned off by default.

Examples:

table.setTableOptions({mergeSeries: {active: true}});
table.display();

table.setTableOptions({
  mergeSeries: {
    active: true,
    mergeOptions: {
      missing: "No Data"
    },
    singleFoi: {
      labelTemplates: {
        headerLabel: "[%offeringName%]"
      }
    },
    multipleFois: {
      labelTemplates: {
        ordinateLabel: "[%offeringName%] [%name%] / [%uomTitle%]"
      }
    }
  }
});
table.display();