Added entryFilter function to SOS.Menu for offerings and observedProperties.
By default, these functions simply pass the offerings entries array, and the observedProperties entries array, as-is to the menu.  However, these functions can be overridden to selectively filter which offerings/observedProperties to show in the menu list boxes.  This can be useful where an offering is of use to an application, but not meant to be selectable by the user.

Furthermore, SOS.Menu has a new function called filterEntries, which takes a menu entries array, and an object that contains either an include array of regexps, or an exclude array of regexps.  The function will then test either the include regexp list, or the exclude regexp list (one or the other, in that order), to the value and label of each menu entry.  This is useful for easily building custom filter functions for the entryFilter function above.

For example:

// Only show offerings from Halley or Rothera
menu.setMenuOptions({
  offerings: {
    entryFilter: function(entries) {
      return menu.filterEntries(entries, {include: [/Halley/, /Rothera/]});
    }
  }
});

// Don't show any pre-averaged offerings
menu.setMenuOptions({
  offerings: {
    entryFilter: function(entries) {
      return menu.filterEntries(entries, {exclude: [/Mean$/]});
    }
  }
});

// Don't show any supplementary engineering observed properties
menu.setMenuOptions({
  observedProperties: {
    entryFilter: function(entries) {
      return menu.filterEntries(entries, {exclude: [/calibration$/, /quality$/]});
    }
  }
});