Mods. to the way FOI ID is parsed in SOS.Map.
In the previous commit, it was incumbent on the user to know in which
element their SOS encoded the FOI ID.  The default setting now is
"auto", which attempts to auto-discover where the FOI ID is held, so
that the user doesn't have to know.  This makes it much more user
friendly, and backwards compatible.  This auto-discovery can be
overridden though, if the user explicitly specifies where to parse the
FOI ID from.  See issue #38.

Dependent on the value of:

SOS.Map.config.featureOfInterestLayer.options.parseFoiIdMethod

the following actions will occur:

* "auto" (the default): the FOI ID will be parsed from the gml:id
attribute, failing that the gml:name element, and failing that the
gml:identifier attribute (this latter option is SOS 2.0 specific).
* "id": the FOI ID will be parsed from the gml:id attribute.
* "name": the FOI ID will be parsed from the gml:name element.
* "identifier": the FOI ID will be parsed from the gml:identifier attribute.